% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FizzBuzz.R
\name{FizzBuzz}
\alias{FizzBuzz}
\title{Play the FizzBuzz game up to n}
\usage{
FizzBuzz(n)
}
\arguments{
\item{n}{A positive integer (the upper limit of the game)}
}
\value{
A character vector of length n containing the FizzBuzz results
}
\description{
Returns a character vector where multiples of 3 are "Fizz",
multiples of 5 are "Buzz", and multiples of both are "Fizz Buzz".
}
